+++ patzilla-ui/navigator/app/document/document-controller.js
@@ -222,13 +222,14 @@ DocumentDetailsController = Marionette.Controller.extend({
 
     get_fulltext: function(document) {
 
-        // Resolve datasource by country
+        // Resolve effective fulltext datasource by country in order of definition
         var datasource_name;
         var country = document.get('@country');
-        _.each(navigatorApp.config.get('system').datasource, function(datasource_info, key) {
+        _.find(navigatorApp.config.get('system').datasources, function(name) {
+            var datasource_info = navigatorApp.config.get('system').datasource[name];
             if (_.contains(datasource_info.fulltext_countries, country)) {
-                datasource_name = key;
-                return;
+                datasource_name = name;
+                return true;
             }
         });