+++ CHANGES.rst
@@ -7,6 +7,7 @@ Development
 ===========
 - [ui] Gradually ramping up ES6
 - [ui] Notify users by MDC Snackbar, ditch NotificationFx
+- [ui] Compute effective full text data source by honouring its definition order
 
 
 2018-11-26 0.161.1